fix linting

+++ examples/janus/inference.py
@@ -44,9 +44,9 @@ def multimodal_understanding(
     tokenizer = vl_chat_processor.tokenizer
 
     pil_images = load_pil_images(conversation)
-    prepare_inputs = vl_chat_processor(
-        conversations=conversation, images=pil_images, force_batchify=True
-    ).to(vl_gpt.dtype)
+    prepare_inputs = vl_chat_processor(conversations=conversation, images=pil_images, force_batchify=True).to(
+        vl_gpt.dtype
+    )
 
     inputs_embeds = vl_gpt.prepare_inputs_embeds(**prepare_inputs)
     st = time()
@@ -81,15 +81,9 @@ def multimodal_understanding(
 
 if __name__ == "__main__":
     parser = argparse.ArgumentParser()
-    parser.add_argument(
-        "--ms_mode", type=int, default=1, help="mindspore mode, 0: graph, 1: pynative"
-    )
-    parser.add_argument(
-        "--image", type=str, default="images/doge.png", help="path to input image"
-    )
-    parser.add_argument(
-        "--question", type=str, default="explain this meme", help="path to input image"
-    )
+    parser.add_argument("--ms_mode", type=int, default=1, help="mindspore mode, 0: graph, 1: pynative")
+    parser.add_argument("--image", type=str, default="images/doge.png", help="path to input image")
+    parser.add_argument("--question", type=str, default="explain this meme", help="path to input image")
     parser.add_argument(
         "--model_path",
         type=str,
@@ -118,9 +112,7 @@ if __name__ == "__main__":
         ms.set_context(jit_config={"jit_level": "O0"}, enable_compile_cache=True)
 
     # specify the path to the model
-    vl_chat_processor: VLChatProcessor = VLChatProcessor.from_pretrained(
-        args.model_path
-    )
+    vl_chat_processor: VLChatProcessor = VLChatProcessor.from_pretrained(args.model_path)
 
     config = AutoConfig.from_pretrained(args.model_path)
     language_config = config.language_config
